Update samsungctl.conf
When id is empty, I explicitly have to authorise remote control on the TV on every start of samsungctl. When id is non-empty, it stores the authorization properly on the TV. Giving it a non-empty value in the example might save others the time it took me to figure out.

See also issue #16, although that has to do with re-authorisation after a reboot (which I dit not observe).